# src/generated/openapi/api/openapi.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- description: Generated server url
  url: http://localhost:8052
paths:
  /api/word/trained:
    post:
      operationId: processTrainedWord
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TrainedWordDto'
        required: true
      responses:
        "200":
          content:
            '*/*':
              schema:
                type: string
          description: OK
      tags:
      - word-controller
      x-content-type: application/json
      x-accepts: '*/*'
  /api/word/save:
    post:
      operationId: addWord
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WordDto'
        required: true
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/WordDto'
          description: OK
      tags:
      - word-controller
      x-content-type: application/json
      x-accepts: '*/*'
...
  /api/dict/reset:
    post:
      operationId: resetWord
      parameters:
      - explode: true
        in: query
        name: wordId
        required: true
        schema:
          format: int64
          type: integer
        style: form
      responses:
        "200":
          description: OK
      tags:
      - dictionary-controller
      x-accepts: application/json
...
  /api/word/find:
    get:
      operationId: findWord
      parameters:
      - explode: true
        in: query
        name: w
        required: true
        schema:
          type: string
        style: form
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/WordDto'
          description: OK
      tags:
      - word-controller
      x-accepts: '*/*'
...
  /api/dict/getWord/{word}:
    get:
      operationId: getWord
      parameters:
      - explode: false
        in: path
        name: word
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/DictWord'
          description: OK
      tags:
      - dictionary-controller
      x-accepts: '*/*'
...
  /api/word/delete:
    delete:
      operationId: deleteWord
      parameters:
      - explode: true
        in: query
        name: id
        required: true
        schema:
          format: int64
          type: integer
        style: form
      responses:
        "200":
          content:
            '*/*':
              schema:
                type: string
          description: OK
      tags:
      - word-controller
      x-accepts: '*/*'
components:
  schemas:
    TrainedWordDto:
      example:
        success: true
        id: 0
        lang: EN
      properties:
        id:
          format: int64
          type: integer
        success:
          type: boolean
        lang:
          enum:
          - EN
          - UA
          type: string
      type: object
    WordDto:
      example:
        ukrainian: ukrainian
        nextTrain: 2000-01-23T04:56:07.000+00:00
        dictWordFreqSum: 2
        aitranslated: true
        description: description
        dictWordId: 5
        userId: 6
        englishCnt: 1
        ukrainianCnt: 5
        examples:
        - examples
        - examples
        lastTrain: 2000-01-23T04:56:07.000+00:00
        english: english
        id: 0
        state: PAUSED
        lang: EN
        rarity: 7
      properties:
        id:
          format: int64
          type: integer
        english:
          type: string
        ukrainian:
          type: string
        description:
          type: string
        userId:
          format: int64
          type: integer
        englishCnt:
          format: int32
          type: integer
        ukrainianCnt:
          format: int32
          type: integer
        lastTrain:
          format: date-time
          type: string
        nextTrain:
          format: date-time
          type: string
        lang:
          enum:
          - EN
          - UA
          type: string
        state:
          enum:
          - PAUSED
          - STAGE_1
          - STAGE_2
          - STAGE_3
          - STAGE_4
          - STAGE_5
          - STAGE_6
          - STAGE_7
          - DONE
          type: string
        dictWordId:
          format: int64
          type: integer
        dictWordFreqSum:
          format: int64
          type: integer
        rarity:
          format: int32
          type: integer
        examples:
          items:
            type: string
          type: array
        aitranslated:
          type: boolean
      type: object
...
    DictTrans:
      example:
        wordText: wordText
        id: 4
      properties:
        id:
          format: int64
          type: integer
        wordText:
          type: string
      type: object
    DictWord:
      example:
        examples:
        - id: 7
          example: example
        - id: 7
          example: example
        translations:
        - wordText: wordText
          id: 4
        - wordText: wordText
          id: 4
        wordText: wordText
        id: 2
        state: 7
        definitions:
        - definitionText: definitionText
          partOfSpeach: partOfSpeach
          id: 2
        - definitionText: definitionText
          partOfSpeach: partOfSpeach
          id: 2
        forms:
        - wordText: wordText
          freq: 3
          id: 9
        - wordText: wordText
          freq: 3
          id: 9
      properties:
        id:
          format: int64
          type: integer
        wordText:
          type: string
        state:
          format: int32
          type: integer
        forms:
          items:
            $ref: '#/components/schemas/DictWordForm'
          type: array
        definitions:
          items:
            $ref: '#/components/schemas/DictWordDefinition'
          type: array
        translations:
          items:
            $ref: '#/components/schemas/DictTrans'
          type: array
        examples:
          items:
            $ref: '#/components/schemas/DictWordExamples'
          type: array
      type: object
    DictWordDefinition:
      example:
        definitionText: definitionText
        partOfSpeach: partOfSpeach
        id: 2
      properties:
        id:
          format: int64
          type: integer
        partOfSpeach:
          type: string
        definitionText:
          type: string
      type: object
    DictWordExamples:
      example:
        id: 7
        example: example
      properties:
        id:
          format: int64
          type: integer
        example:
          type: string
      type: object
    DictWordForm:
      example:
        wordText: wordText
        freq: 3
        id: 9
      properties:
        id:
          format: int64
          type: integer
        wordText:
          type: string
        freq:
          format: int64
          type: integer
      type: object
...