# src/main/java/org/sav/fornas/cards/controller/WordController.java
package org.sav.fornas.cards.controller;

import jakarta.servlet.http.HttpServletRequest;
import lombok.RequiredArgsConstructor;
import lombok.extern.slf4j.Slf4j;
import org.sav.fornas.cards.client.cardsback.model.*;
import org.sav.fornas.cards.security.TokenService;
import org.sav.fornas.cards.service.DictionaryService;
import org.sav.fornas.cards.service.WordService;
import org.springframework.security.core.annotation.AuthenticationPrincipal;
import org.springframework.security.oauth2.core.oidc.user.OidcUser;
import org.springframework.stereotype.Controller;
import org.springframework.ui.Model;
import org.springframework.web.bind.WebDataBinder;
import org.springframework.web.bind.annotation.*;

import java.beans.PropertyEditorSupport;
import java.time.OffsetDateTime;
import java.util.List;

@Controller
@Slf4j
@RequiredArgsConstructor
public class WordController {

	private final WordService wordService;
	private final DictionaryService dictionaryService;
	private final TokenService tokenService;

	@GetMapping("/words")
	public String viewInfo(
			@RequestParam(defaultValue = "0") int page,
			@RequestParam(defaultValue = "20") int size,
			@RequestParam(defaultValue = "") String state,
			Model model) {

		WordsPageDtoWordDto words = wordService.getWordsByUser(page, size, state);
		model.addAttribute("words", words);
		return "words";
	}

	@GetMapping("/edit")
	public String showForm(Model model, @RequestParam(name = "w", defaultValue = "") String w) {
		WordDto word = wordService.findWord(w.toLowerCase());
		if (word.getState() == null) {
			word.setState(WordDto.StateEnum.STAGE_1);
		}
		model.addAttribute("word", word);
		return "word-form";
	}

	@GetMapping("/reset")
	public String resetWord(@RequestParam(name = "id", defaultValue = "") Long id, @RequestParam(name = "w", defaultValue = "") String w) {
		dictionaryService.resetWord(id);
		return "redirect:/edit?w=" + w;
	}

	@GetMapping("/random")
	public String randomWord(@AuthenticationPrincipal OidcUser user, Model model, HttpServletRequest request) {

		Long key = (Long) user.getClaims().get("userId");
		log.debug(">>> key={}", key);
		boolean isUpdating = dictionaryService.isUpdating(key);

		log.debug(">>> isUpdating={}", isUpdating);
		List<WordDto> words = dictionaryService.getWords(key);
		if(words.size() <= 3 && !isUpdating) {
			log.debug(">>> initiating update");

			String token = tokenService.getAccessToken();

			dictionaryService.getNewWordsAsync(key, token)
					.thenRun(() -> log.debug(">>> async update finished for key {}", key));

		}

		WordDto word = !words.isEmpty() ? words.removeFirst() : null;
		model.addAttribute("word", word);
		model.addAttribute("randomListSize", words.size());
		model.addAttribute("returnUrl", request.getServletPath());
		return "word-form";
	}

	@GetMapping("/randomOne")
	public String randomOneWord(Model model, HttpServletRequest request) {

		WordDto word = dictionaryService.getNewWord();
		model.addAttribute("word", word);
		model.addAttribute("returnUrl", request.getServletPath());
		return "word-form";
	}

	@GetMapping("/pick5Paused")
	public String pick10Paused() {
		wordService.pick5Paused();
		return "redirect:/statistic";
	}

	@GetMapping("/add")
	public String addWord() {
		return "add-word";
	}

	@PostMapping("/save")
	public String saveWord(@ModelAttribute WordDto wordDto) {

		log.debug(">>> saveWord={}", wordDto);
		WordDto newWord = wordService.saveWord(wordDto);
		return "redirect:/edit?w=" + newWord.getEnglish();
	}

	@PostMapping("/setSkipped")
	public String setSkipped(@RequestParam Long dictWordId, @RequestParam String returnUrl) {
		wordService.setMark(dictWordId, "SKIP");
		return "redirect:" + returnUrl;
	}

	@PostMapping("/setKnown")
	public String setKnown(@RequestParam Long dictWordId, @RequestParam String returnUrl) {
		wordService.setMark(dictWordId, "KNOWN");
		return "redirect:" + returnUrl;
	}

	@GetMapping("/delete")
	public String deleteWord(@RequestParam("id") Long id) {
		wordService.deleteWord(id);
		return "redirect:/words";
	}

	@GetMapping("/statistic")
	public String statistic(Model model) {
		StatisticDto statisticDto = wordService.getStatistics();
		log.debug(">>> statistic={}", statisticDto);
		model.addAttribute("statistics", statisticDto);
		return "statistic";
	}

	@GetMapping("/train")
	public String train(Model model) {
		WordDto word = wordService.getWord();
		if(word == null){
			return "redirect:/add";
		}
		StateLimitDto stateLimit = wordService.getStateLimit(word.getState().getValue());
		int minCnt = word.getEnglishCnt() + word.getUkrainianCnt();
		double progressPercent = (double) (minCnt + 1) * 100 / (stateLimit.getAttempt() * 2);

		model.addAttribute("progressPercent", progressPercent);
		model.addAttribute("word", word);
		model.addAttribute("stateColor", stateLimit.getColor());
		model.addAttribute("mode", "train");
		return "train";
	}

	@GetMapping("/retro")
	public String retro(@AuthenticationPrincipal OidcUser user, Model model) {
		Long key = (Long) user.getClaims().get("userId");
		WordDto word = wordService.getRetroWord(key);
		if(word == null){
			return "redirect:/add";
		}
		model.addAttribute("word", word);
		model.addAttribute("mode", "retro");
		return "train";
	}

	@GetMapping("/trained")
	public String setTrained(@ModelAttribute TrainedWordDto trainedWordDto) {
		log.debug(">>> trainedWord={}", trainedWordDto);
		wordService.setTrained(trainedWordDto);
		return "redirect:/train";
	}



	@InitBinder
	public void initBinder(WebDataBinder binder) {
		binder.registerCustomEditor(OffsetDateTime.class, new PropertyEditorSupport() {
			@Override
			public void setAsText(String text) {
				if (text == null || text.isEmpty()) {
					setValue(null);
				} else {
					setValue(OffsetDateTime.parse(text));
				}
			}
		});
	}
}